MovieRecommender/requirements.txt
flask
flask-sqlalchemy
flask-user==1.0.2.2
bcrypt==4.0.1
tensorflow==2.10.1
blinker
